Paste output from chat gpt
Prompt was:

Can you build me a simple website that starts with a play button. When the play button is pressed, 1 random image from a list of 10 images. A text box is also shown and an enter button. When the enter button is pressed, the site checks if the text in the text box matches the name of the image. If it's a match, add 1 to the score. Then load the next image at random. After 10 images, show a final score page with a button to restart. Include all HTML and JavaScript